+++ ddd/logic/encodage_des_notes/soumission/test/use_case/read/test_get_progression_generale_encodage_service.py
@@ -29,8 +29,8 @@ from django.test import SimpleTestCase
 
 from ddd.logic.encodage_des_notes.soumission.commands import GetProgressionGeneraleCommand
 from ddd.logic.encodage_des_notes.soumission.dtos import SignaletiqueEtudiantDTO, AttributionEnseignantDTO
-from ddd.logic.encodage_des_notes.tests.factory.note_etudiant import NoteDejaSoumise
-from ddd.logic.encodage_des_notes.tests.factory.note_etudiant import NoteManquanteEtudiantFactory
+from ddd.logic.encodage_des_notes.soumission.test.factory import NoteDejaSoumise
+from ddd.logic.encodage_des_notes.soumission.test.factory import NoteManquanteEtudiantFactory
 from infrastructure.encodage_de_notes.shared_kernel.service.in_memory.attribution_enseignant import \
     AttributionEnseignantTranslatorInMemory
 from infrastructure.encodage_de_notes.shared_kernel.service.in_memory.periode_encodage_notes import \
