+++ ddd/logic/encodage_des_notes/soumission/test/use_case/write/test_soumettre_note.py
@@ -24,7 +24,6 @@
 #
 ##############################################################################
 import datetime
-from datetime import date
 from unittest import mock
 
 import attr
@@ -35,9 +34,9 @@ from ddd.logic.encodage_des_notes.shared_kernel.dtos import DateDTO, PeriodeEnco
 from ddd.logic.encodage_des_notes.soumission.commands import SoumettreNoteCommand
 from ddd.logic.encodage_des_notes.soumission.domain.validator.exceptions import PasResponsableDeNotesException
 from ddd.logic.encodage_des_notes.shared_kernel.validator.exceptions import PeriodeEncodageNotesFermeeException
-from ddd.logic.encodage_des_notes.tests.factory.note_etudiant import NoteChiffreEtudiantFactory, \
+from ddd.logic.encodage_des_notes.soumission.test.factory import NoteChiffreEtudiantFactory, \
     NoteManquanteEtudiantFactory
-from ddd.logic.encodage_des_notes.tests.factory.responsable_de_notes import \
+from ddd.logic.encodage_des_notes.soumission.test.factory.responsable_de_notes import \
     ResponsableDeNotesLDROI1001Annee2020Factory
 from infrastructure.encodage_de_notes.shared_kernel.service.in_memory.periode_encodage_notes import \
     PeriodeEncodageNotesTranslatorInMemory
